ep05- Lets get hooked
