--- App.js
@@ -1,587 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-import logo from "./Assests/logo.png";
-import shopping_cart from "./Assests/shopping_cart.svg";
-import search_icon from "./Assests/search_icon.svg";
-import Pizzaimage from "./Assests/Pizza image.png";
-
-const Header = () => {
-  return (
-    <div className="header">
-      <div className="logo-container">
-        <img className="logo" src={logo}></img>
-      </div>
-
-      <ul className="nav-items">
-        <li>Home</li>
-        <li>About</li>
-        <li>Contact</li>
-      </ul>
-
-      <ul className="nav-items">
-        <li>
-          <img src={shopping_cart}></img>
-        </li>
-        <li>Profile</li>
-      </ul>
-    </div>
-  );
-};
-
-const Restaurantcard = (props) => {
-  const { resData } = props;
-  const { cloudinaryImageId, cuisines, name, costForTwo } = resData?.info;
-
-  return (
-    <div className="res-card">
-      <img
-        className="pizza-image"
-        src={
-          "https://media-assets.swiggy.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_660/" +
-          cloudinaryImageId
-        }
-      ></img>
-      <div>
-        <p>{name}</p>
-        <p>{cuisines.join(", ")}</p>
-        <p>{costForTwo}</p>
-      </div>
-    </div>
-  );
-};
-const resList = [
-  {
-    info: {
-      id: "55473",
-      name: "Pizza Hut",
-      cloudinaryImageId:
-        "RX_THUMBNAIL/IMAGES/VENDOR/2024/7/16/d281ae33-578d-482f-a4e2-8cf0d8ee3d58_55473.jpg",
-      locality: "Sector 10",
-      areaName: "Rohini",
-      costForTwo: "₹350 for two",
-      cuisines: ["Pizzas"],
-      avgRating: 4.1,
-      parentId: "721",
-      avgRatingString: "4.1",
-      totalRatingsString: "7.5K+",
-      sla: {
-        deliveryTime: 26,
-        lastMileTravel: 2.1,
-        serviceability: "SERVICEABLE",
-        slaString: "25-30 mins",
-        lastMileTravelString: "2.1 km",
-        iconType: "ICON_TYPE_EMPTY",
-      },
-      availability: {
-        nextCloseTime: "2024-12-29 04:00:00",
-        opened: true,
-      },
-      badges: {
-        imageBadges: [
-          {
-            imageId: "Rxawards/_CATEGORY-Pizza.png",
-            description: "Delivery!",
-          },
-        ],
-      },
-      isOpen: true,
-      type: "F",
-      badgesV2: {
-        entityBadges: {
-          imageBased: {
-            badgeObject: [
-              {
-                attributes: {
-                  description: "Delivery!",
-                  imageId: "Rxawards/_CATEGORY-Pizza.png",
-                },
-              },
-            ],
-          },
-          textBased: {},
-          textExtendedBadges: {},
-        },
-      },
-      aggregatedDiscountInfoV3: {
-        header: "50% OFF",
-        subHeader: "UPTO ₹90",
-      },
-      differentiatedUi: {
-        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-        differentiatedUiMediaDetails: {
-          lottie: {},
-          video: {},
-        },
-      },
-      reviewsSummary: {},
-      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-      restaurantOfferPresentationInfo: {},
-      externalRatings: {
-        aggregatedRating: {
-          rating: "--",
-        },
-      },
-      ratingsDisplayPreference: "RATINGS_DISPLAY_PREFERENCE_SHOW_SWIGGY",
-    },
-    analytics: {
-      context: "seo-data-61d4333a-f721-4d4a-a599-6b5879bcba12",
-    },
-    cta: {
-      link: "https://www.swiggy.com/city/delhi/pizza-hut-sector-10-rohini-rest55473",
-      type: "WEBLINK",
-    },
-  },
-  {
-    info: {
-      id: "606512",
-      name: "Chinese Wok",
-      cloudinaryImageId: "e0839ff574213e6f35b3899ebf1fc597",
-      locality: "Sector 11",
-      areaName: "Rohini",
-      costForTwo: "₹250 for two",
-      cuisines: ["Chinese", "Asian", "Tibetan", "Desserts"],
-      avgRating: 4.1,
-      parentId: "61955",
-      avgRatingString: "4.1",
-      totalRatingsString: "1.5K+",
-      sla: {
-        deliveryTime: 37,
-        lastMileTravel: 4.7,
-        serviceability: "SERVICEABLE",
-        slaString: "35-40 mins",
-        lastMileTravelString: "4.7 km",
-        iconType: "ICON_TYPE_EMPTY",
-      },
-      availability: {
-        nextCloseTime: "2024-12-29 03:00:00",
-        opened: true,
-      },
-      badges: {
-        imageBadges: [
-          {
-            imageId: "Green%20Dot%20Awards/Best%20In%20Veg%20Noodles.png",
-            description: "Delivery!",
-          },
-        ],
-      },
-      isOpen: true,
-      type: "F",
-      badgesV2: {
-        entityBadges: {
-          imageBased: {
-            badgeObject: [
-              {
-                attributes: {
-                  description: "Delivery!",
-                  imageId: "Green%20Dot%20Awards/Best%20In%20Veg%20Noodles.png",
-                },
-              },
-            ],
-          },
-          textBased: {},
-          textExtendedBadges: {},
-        },
-      },
-      aggregatedDiscountInfoV3: {
-        header: "ITEMS",
-        subHeader: "AT ₹149",
-      },
-      differentiatedUi: {
-        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-        differentiatedUiMediaDetails: {
-          lottie: {},
-          video: {},
-        },
-      },
-      reviewsSummary: {},
-      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-      restaurantOfferPresentationInfo: {},
-      externalRatings: {
-        aggregatedRating: {
-          rating: "--",
-        },
-      },
-      ratingsDisplayPreference: "RATINGS_DISPLAY_PREFERENCE_SHOW_SWIGGY",
-    },
-    analytics: {
-      context: "seo-data-61d4333a-f721-4d4a-a599-6b5879bcba12",
-    },
-    cta: {
-      link: "https://www.swiggy.com/city/delhi/chinese-wok-sector-11-rohini-rest606512",
-      type: "WEBLINK",
-    },
-  },
-  {
-    info: {
-      id: "498382",
-      name: "Burger King",
-      cloudinaryImageId:
-        "RX_THUMBNAIL/IMAGES/VENDOR/2024/8/9/64339639-9425-4d7a-807f-6fca82056ec3_498382.jpg",
-      locality: "M2K Mall",
-      areaName: "Rohini",
-      costForTwo: "₹350 for two",
-      cuisines: ["Burgers", "American"],
-      avgRating: 4.5,
-      parentId: "166",
-      avgRatingString: "4.5",
-      totalRatingsString: "5.4K+",
-      sla: {
-        deliveryTime: 24,
-        lastMileTravel: 2,
-        serviceability: "SERVICEABLE",
-        slaString: "20-25 mins",
-        lastMileTravelString: "2.0 km",
-        iconType: "ICON_TYPE_EMPTY",
-      },
-      availability: {
-        nextCloseTime: "2024-12-29 03:00:00",
-        opened: true,
-      },
-      badges: {
-        imageBadges: [
-          {
-            imageId: "bolt/Bolt%20Listing%20badge@3x.png",
-            description: "bolt!",
-          },
-          {
-            imageId: "Rxawards/_CATEGORY-Burger.png",
-            description: "Delivery!",
-          },
-        ],
-      },
-      isOpen: true,
-      type: "F",
-      badgesV2: {
-        entityBadges: {
-          imageBased: {
-            badgeObject: [
-              {
-                attributes: {
-                  description: "bolt!",
-                  imageId: "bolt/Bolt%20Listing%20badge@3x.png",
-                },
-              },
-              {
-                attributes: {
-                  description: "Delivery!",
-                  imageId: "Rxawards/_CATEGORY-Burger.png",
-                },
-              },
-            ],
-          },
-          textBased: {},
-          textExtendedBadges: {},
-        },
-      },
-      aggregatedDiscountInfoV3: {
-        header: "60% OFF",
-        subHeader: "UPTO ₹120",
-      },
-      differentiatedUi: {
-        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-        differentiatedUiMediaDetails: {
-          lottie: {},
-          video: {},
-        },
-      },
-      reviewsSummary: {},
-      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-      restaurantOfferPresentationInfo: {},
-      externalRatings: {
-        aggregatedRating: {
-          rating: "--",
-        },
-      },
-      ratingsDisplayPreference: "RATINGS_DISPLAY_PREFERENCE_SHOW_SWIGGY",
-    },
-    analytics: {
-      context: "seo-data-61d4333a-f721-4d4a-a599-6b5879bcba12",
-    },
-    cta: {
-      link: "https://www.swiggy.com/city/delhi/burger-king-m2k-mall-rohini-rest498382",
-      type: "WEBLINK",
-    },
-  },
-  {
-    info: {
-      id: "622616",
-      name: "Domino's Pizza",
-      cloudinaryImageId:
-        "RX_THUMBNAIL/IMAGES/VENDOR/2024/11/29/5c5d1d92-c9b5-42ce-9e74-5b3c2cb960e4_622616.jpg",
-      locality: "3rd Sector",
-      areaName: "Rohini",
-      costForTwo: "₹400 for two",
-      cuisines: ["Pizzas", "Italian", "Pastas", "Desserts"],
-      avgRating: 4.5,
-      parentId: "2456",
-      avgRatingString: "4.5",
-      totalRatingsString: "1.3K+",
-      sla: {
-        deliveryTime: 30,
-        lastMileTravel: 0.8,
-        serviceability: "SERVICEABLE",
-        slaString: "25-30 mins",
-        lastMileTravelString: "0.8 km",
-        iconType: "ICON_TYPE_EMPTY",
-      },
-      availability: {
-        nextCloseTime: "2024-12-29 02:59:00",
-        opened: true,
-      },
-      badges: {
-        imageBadges: [
-          {
-            imageId: "bolt/Bolt%20Listing%20badge@3x.png",
-            description: "bolt!",
-          },
-          {
-            imageId: "Rxawards/_CATEGORY-Pizza.png",
-            description: "Delivery!",
-          },
-        ],
-      },
-      isOpen: true,
-      type: "F",
-      badgesV2: {
-        entityBadges: {
-          imageBased: {
-            badgeObject: [
-              {
-                attributes: {
-                  description: "bolt!",
-                  imageId: "bolt/Bolt%20Listing%20badge@3x.png",
-                },
-              },
-              {
-                attributes: {
-                  description: "Delivery!",
-                  imageId: "Rxawards/_CATEGORY-Pizza.png",
-                },
-              },
-            ],
-          },
-          textBased: {},
-          textExtendedBadges: {},
-        },
-      },
-      aggregatedDiscountInfoV3: {
-        header: "ITEMS",
-        subHeader: "AT ₹89",
-      },
-      differentiatedUi: {
-        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-        differentiatedUiMediaDetails: {
-          lottie: {},
-          video: {},
-        },
-      },
-      reviewsSummary: {},
-      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-      restaurantOfferPresentationInfo: {},
-      externalRatings: {
-        aggregatedRating: {
-          rating: "--",
-        },
-      },
-      ratingsDisplayPreference: "RATINGS_DISPLAY_PREFERENCE_SHOW_SWIGGY",
-    },
-    analytics: {
-      context: "seo-data-61d4333a-f721-4d4a-a599-6b5879bcba12",
-    },
-    cta: {
-      link: "https://www.swiggy.com/city/delhi/dominos-pizza-3rd-sector-rohini-rest622616",
-      type: "WEBLINK",
-    },
-  },
-  {
-    info: {
-      id: "614570",
-      name: "Subway",
-      cloudinaryImageId:
-        "RX_THUMBNAIL/IMAGES/VENDOR/2024/12/9/28bacc3b-d4dc-482a-86e3-cab0547a9c78_614570.JPG",
-      locality: "Sector 10",
-      areaName: "Rohini",
-      costForTwo: "₹350 for two",
-      cuisines: ["sandwich", "Salads", "wrap", "Healthy Food"],
-      avgRating: 4,
-      parentId: "2",
-      avgRatingString: "4.0",
-      totalRatingsString: "1.4K+",
-      sla: {
-        deliveryTime: 18,
-        lastMileTravel: 2.1,
-        serviceability: "SERVICEABLE",
-        slaString: "15-20 mins",
-        lastMileTravelString: "2.1 km",
-        iconType: "ICON_TYPE_EMPTY",
-      },
-      availability: {
-        nextCloseTime: "2024-12-29 01:00:00",
-        opened: true,
-      },
-      badges: {
-        imageBadges: [
-          {
-            imageId: "Green%20Dot%20Awards/Best%20In%20Veg%20Salad.png",
-            description: "Delivery!",
-          },
-        ],
-      },
-      isOpen: true,
-      type: "F",
-      badgesV2: {
-        entityBadges: {
-          imageBased: {
-            badgeObject: [
-              {
-                attributes: {
-                  description: "Delivery!",
-                  imageId: "Green%20Dot%20Awards/Best%20In%20Veg%20Salad.png",
-                },
-              },
-            ],
-          },
-          textBased: {},
-          textExtendedBadges: {},
-        },
-      },
-      aggregatedDiscountInfoV3: {
-        header: "ITEMS",
-        subHeader: "AT ₹119",
-      },
-      differentiatedUi: {
-        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-        differentiatedUiMediaDetails: {
-          lottie: {},
-          video: {},
-        },
-      },
-      reviewsSummary: {},
-      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-      restaurantOfferPresentationInfo: {},
-      externalRatings: {
-        aggregatedRating: {
-          rating: "--",
-        },
-      },
-      ratingsDisplayPreference: "RATINGS_DISPLAY_PREFERENCE_SHOW_SWIGGY",
-    },
-    analytics: {
-      context: "seo-data-61d4333a-f721-4d4a-a599-6b5879bcba12",
-    },
-    cta: {
-      link: "https://www.swiggy.com/city/delhi/subway-sector-10-rohini-rest614570",
-      type: "WEBLINK",
-    },
-  },
-  {
-    info: {
-      id: "225317",
-      name: "La Pino'z Pizza",
-      cloudinaryImageId: "aizb6uyem2jb1hcrhxac",
-      locality: "Rohini",
-      areaName: "New Friends Colony",
-      costForTwo: "₹500 for two",
-      cuisines: ["Pizzas", "Pastas", "Italian", "Desserts", "Beverages"],
-      avgRating: 4.3,
-      parentId: "4961",
-      avgRatingString: "4.3",
-      totalRatingsString: "17K+",
-      sla: {
-        deliveryTime: 27,
-        lastMileTravel: 4,
-        serviceability: "SERVICEABLE",
-        slaString: "25-30 mins",
-        lastMileTravelString: "4.0 km",
-        iconType: "ICON_TYPE_EMPTY",
-      },
-      availability: {
-        nextCloseTime: "2024-12-29 05:00:00",
-        opened: true,
-      },
-      badges: {
-        imageBadges: [
-          {
-            imageId: "Green%20Dot%20Awards/Best%20In%20Veg%20Pizza.png",
-            description: "Delivery!",
-          },
-        ],
-      },
-      isOpen: true,
-      type: "F",
-      badgesV2: {
-        entityBadges: {
-          imageBased: {
-            badgeObject: [
-              {
-                attributes: {
-                  description: "Delivery!",
-                  imageId: "Green%20Dot%20Awards/Best%20In%20Veg%20Pizza.png",
-                },
-              },
-            ],
-          },
-          textBased: {},
-          textExtendedBadges: {},
-        },
-      },
-      aggregatedDiscountInfoV3: {
-        header: "50% OFF",
-        subHeader: "UPTO ₹90",
-      },
-      differentiatedUi: {
-        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
-        differentiatedUiMediaDetails: {
-          lottie: {},
-          video: {},
-        },
-      },
-      reviewsSummary: {},
-      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
-      restaurantOfferPresentationInfo: {},
-      externalRatings: {
-        aggregatedRating: {
-          rating: "3.9",
-          ratingCount: "245",
-        },
-        source: "GOOGLE",
-        sourceIconImageId: "v1704440323/google_ratings/rating_google_tag",
-      },
-      ratingsDisplayPreference: "RATINGS_DISPLAY_PREFERENCE_SHOW_SWIGGY",
-    },
-    analytics: {
-      context: "seo-data-61d4333a-f721-4d4a-a599-6b5879bcba12",
-    },
-    cta: {
-      link: "https://www.swiggy.com/city/delhi/la-pinoz-pizza-rohini-new-friends-colony-rest225317",
-      type: "WEBLINK",
-    },
-  },
-];
-
-const Body = () => {
-  return (
-    <div className="body">
-      <div className="search">
-        <input placeholder="Order Something !"></input>
-        <img src={search_icon}></img>
-      </div>
-      <div className="res-container">
-        {resList.map((data) => (
-          <Restaurantcard key={data.info.id} resData={data} />
-        ))}
-      </div>
-    </div>
-  );
-};
-const AppLayout = () => {
-  return (
-    <div className="app">
-      <Header />
-      <Body />
-    </div>
-  );
-};
-const root = ReactDOM.createRoot(document.getElementById("root"));
-
-root.render(<AppLayout></AppLayout>);
