+++ index.css
@@ -16,6 +16,8 @@
   padding-left: 1.5rem;
   padding-right: 1.5rem;
   box-shadow: rgba(173, 168, 168, 0.437) 0px 2px 8px 0px;
+  color: #888684;
+  font-weight: 500;
 }
 
 .logo {
@@ -27,6 +29,38 @@
   gap: 13px;
   list-style-type: none;
   align-items: center;
+}
+.hero {
+  display: flex;
+  width: 80%;
+  margin-left: auto;
+  margin-right: auto;
+  margin-top: 3rem;
+  margin-bottom: 3rem;
+}
+.hero_text,
+.hero_image {
+  flex: 1;
+}
+.hero_text {
+  display: flex;
+  flex-direction: column;
+  gap: 3rem;
+}
+.hero_text > h1 {
+  font-weight: 700;
+  font-size: 80;
+  line-height: 5rem;
+  position: relative;
+}
+
+.hero_text > p {
+  color: #888684;
+}
+
+.hero_image > img {
+  max-height: 70vh;
+  border-radius: 20px;
 }
 
 .body {
@@ -55,6 +89,18 @@
   border: none;
   padding: 5px;
   width: 100%;
+}
+
+.filter-btn {
+  margin: 1rem 1rem 1.5rem 1rem;
+  padding: 0.5rem;
+  outline: none;
+  border: none;
+  border: 1px solid #e55427;
+  border-radius: 10px;
+  background-color: #e55427;
+  cursor: pointer;
+  color: #ffffff;
 }
 
 .res-container {
