+++ index.html
@@ -13,8 +13,8 @@
     rel="stylesheet"
   />
   <body>
-    <div id="root">Hello Shivam</div>
+    <div id="root">Not rendered</div>
 
-    <script type="module" src="App.js"></script>
+    <script type="module" src="src/App.js"></script>
   </body>
 </html>
