+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "new-folder",
+  "name": "namastey-react",
   "version": "1.0.0",
   "description": "This is namastey react",
