+++ src/Assests/assets.js
@@ -1,0 +1,13 @@
+import Logo from "./logo.png";
+import search_icon from "./search_icon.svg";
+import shopping_cart from "../Assests/shopping_cart.svg";
+import pizzaimage from "../Assests/pizza-image.png";
+const AssetPath = {
+  Logo,
+  search_icon,
+  shopping_cart,
+  pizzaimage,
+};
+
+export default AssetPath;
+export { Logo, search_icon, shopping_cart };
